Split config up
