+++ .flake8
@@ -16,3 +16,4 @@ per-file-ignores =
     # SQLAlchemy requires != and == be used when comparing to None and
     # Booleans so ignore these warnings in models and other query heavy modules
     dashboard/models/*.py:E711,E712
+    config/__init__.py:F403,F401
